# docs/swagger.yaml
definitions:
...
  dbmodels.LimitTransaction:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      featureProduct:
        type: string
      id:
        type: integer
      levelMerchant:
        type: string
      limitAmount:
        type: integer
      limitFreq:
        type: integer
      minLimitAmount:
        type: integer
      timeFrame:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      userCategoryId:
        type: string
    type: object
  dbmodels.LogUpgradeFds:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
      req:
        type: string
      res:
        type: string
      retryAt:
        type: string
      retryBy:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  dbmodels.SettlementConfig:
    properties:
      bankCode:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      settlementType:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
  dto.ReqInstructionListDto:
    properties:
      description:
        type: string
      id:
        type: integer
      limit:
        type: integer
      logo:
        type: string
      page:
        type: integer
      sequence:
        type: integer
      title:
        type: string
    type: object
  dto.ReqLimitTransactionDto:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      featureProduct:
        type: string
      id:
        type: integer
      levelMerchant:
        type: string
      limit:
        type: integer
      limitAmount:
        type: integer
      limitFreq:
        type: integer
      minLimitAmount:
        type: integer
      page:
        type: integer
      timeFrame:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      userCategoryId:
        type: string
    type: object
...
  dto.ReqLogUpgradeFdsDto:
    properties:
      limit:
        type: integer
      page:
        type: integer
      phoneNumber:
        type: string
      status:
        type: string
    type: object
...
  dto.ReqSettlementConfigDto:
    properties:
      bankCode:
        description: ID             int64  `json:"id"`
        type: string
      limit:
        type: integer
      page:
        type: integer
      settlementType:
        type: string
      status:
        type: string
      user:
        type: string
    type: object
...
  dto.ResInstructionListDto:
    properties:
      description:
        type: string
      id:
        type: integer
      logo:
        type: string
      sequence:
        type: integer
      title:
        type: string
    type: object
...
  models.Response:
    properties:
      contents: {}
      data: {}
      errCode:
        type: string
      errDesc:
        type: string
      totalData:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /rosego/v.0.1/bank-list:
    post:
      consumes:
      - application/json
      description: SettlementConfigController Save/EDit
      operationId: SettlementConfigController - Save/Edit
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqSettlementConfigDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: SettlementConfigController - Save / Edit
  /rosego/v.0.1/bank-list/filter:
    post:
      consumes:
      - application/json
      description: Settlement Config Filter Paging
      operationId: SettlementConfig - Filter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqSettlementConfigDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dbmodels.SettlementConfig'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: SettlementConfig - Filter
...
  /rosego/v.0.1/instruction-list:
    post:
      consumes:
      - application/json
      description: Instruction List Save
      operationId: Instruction List - Save
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqInstructionListDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Instruction List - Save
  /rosego/v.0.1/instruction-list/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Instruction List
      operationId: Instruction List - Delete
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the order to be deleted
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Instruction List - Delete
  /rosego/v.0.1/instruction-list/filter:
    post:
      consumes:
      - application/json
      description: Instruction List Filter
      operationId: Instruction List - Filter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqInstructionListDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dto.ResInstructionListDto'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Instruction List - Filter
  /rosego/v.0.1/limit-transaction:
    post:
      consumes:
      - application/json
      description: Limit Transaction Save/EDit
      operationId: Limit Transaction - Save/Edit
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqLimitTransactionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Limit Transaction - Save / Edit
  /rosego/v.0.1/limit-transaction/filter:
    post:
      consumes:
      - application/json
      description: Limit Transaction Filter
      operationId: Limit Transaction - Filter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqLimitTransactionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    $ref: '#/definitions/dbmodels.LimitTransaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Limit Transaction - Filter
...
  /rosego/v.0.1/upgrade-fds:
    post:
      consumes:
      - application/json
      description: Upgrade FDS Save/EDit
      operationId: Upgrade FDS - Save/Edit
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dbmodels.LogUpgradeFds'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Upgrade FDS - Save / Edit
  /rosego/v.0.1/upgrade-fds/filter:
    post:
      consumes:
      - application/json
      description: Upgrade FDS Filter Paging
      operationId: Upgrade FDS - Filter
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReqLogUpgradeFdsDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                contents:
                  items:
                    items:
                      $ref: '#/definitions/dbmodels.LogUpgradeFds'
                    type: array
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Upgrade FDS - Filter
...
swagger: "2.0"